+++ src/api/tmc.ts
@@ -588,7 +588,7 @@ export default class TMC {
                 args: ["submit", "--submission-path", exercisePath, "--submission-url", submitUrl],
                 core: true,
                 onStdout: (res) =>
-                    progressCallback?.(50 * res["percent-done"], res.message || undefined),
+                    progressCallback?.(100 * res["percent-done"], res.message || undefined),
             },
             createIs<SubmissionStatusReport>(),
         ).then((res) => res.map((r) => r.data));